# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_rekall_backend_internal_interfaces_http_dto.ChatMessageListResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ChatMessageListPayload'
      success:
        example: true
        type: boolean
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.CreateMeetingRequest:
    properties:
      scope_id:
        type: string
      scope_type:
        type: string
      title:
        type: string
      type:
        enum:
        - open
        - private
        type: string
    required:
    - type
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorBody'
      success:
        example: false
        type: boolean
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.MeetingResponseEnvelope:
    properties:
      data:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.MeetingResponse'
      success:
        example: true
        type: boolean
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.MessageResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.MsgObj'
      success:
        example: true
        type: boolean
    type: object
...
  github_com_rekall_backend_internal_interfaces_http_dto.WSTicketResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.WSTicketPayload'
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: engineering@rekall.io
    name: Rekall Engineering
  description: |-
    Rekall is a call intelligence platform. Record, transcribe, and analyse voice and video calls.
    This API powers the Rekall web application and is available for direct integration.

    ## Authentication
    Protected endpoints require a JWT access token obtained from `POST /api/v1/auth/login`.
    Pass the token in the `Authorization` header as `Bearer <token>`.
    The refresh token is transported exclusively via an `HttpOnly` cookie set by the server.

    ## Response Envelope
    All responses are wrapped in a standard envelope:
    - Success: `{ "success": true, "data": <T>, "meta": <pagination|null> }`
    - Error:   `{ "success": false, "error": { "code": "ERROR_CODE", "message": "Human readable message", "details": <null|object> } }`

    ## Error Codes
    Every error response includes a machine-readable `code` field. Common codes:
    `INVALID_CREDENTIALS`, `EMAIL_NOT_VERIFIED`, `EMAIL_ALREADY_REGISTERED`,
    `INVALID_OR_EXPIRED_TOKEN`, `INVALID_REFRESH_TOKEN`, `MISSING_TOKEN`, `INVALID_TOKEN`,
    `CALL_NOT_FOUND`, `USER_NOT_FOUND`, `SLUG_ALREADY_TAKEN`, `ALREADY_A_MEMBER`,
    `CANNOT_REMOVE_OWNER`, `INVITATION_EMAIL_MISMATCH`, `DEPARTMENT_NOT_FOUND`, `FORBIDDEN`
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Rekall API
  version: 1.0.0
paths:
...
  /api/v1/meetings:
    post:
      consumes:
      - application/json
      description: Creates a new meeting room. Open meetings allow any authenticated
        user to join directly; private meetings restrict access to org/department
        members.
      parameters:
      - description: Meeting creation parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.CreateMeetingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Meeting created
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.MeetingResponseEnvelope'
        "400":
          description: Invalid request body or host meeting limit reached
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a meeting
      tags:
      - Meetings
  /api/v1/meetings/{code}:
    delete:
      description: Ends an active meeting. Only the host may call this. Marks all
        participants as left and sets ended_at.
      parameters:
      - description: Meeting code
        example: abc-defg-hij
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Meeting ended
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.MessageResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "403":
          description: Caller is not the host
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "404":
          description: Meeting not found
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: End a meeting
      tags:
      - Meetings
    get:
      description: Returns a single meeting by its short code (e.g. abc-defg-hij).
        Does not require authentication.
      parameters:
      - description: Meeting code
        example: abc-defg-hij
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Meeting record
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.MeetingResponseEnvelope'
        "404":
          description: Meeting not found
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      summary: Get a meeting
      tags:
      - Meetings
  /api/v1/meetings/{code}/messages:
    get:
      description: Returns the chat history for a meeting, ordered by sent_at ascending.
        Supports cursor-based pagination via the `before` query parameter (messages
        strictly older than the cursor are returned). Accessible to the host and to
        anyone who has ever been an admitted participant.
      parameters:
      - description: Meeting code
        example: abc-defg-hij
        in: path
        name: code
        required: true
        type: string
      - description: RFC3339 cursor — returns messages strictly older
        example: "2026-04-23T14:03:17Z"
        in: query
        name: before
        type: string
      - default: 50
        description: Page size (default 50, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Chat history page
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ChatMessageListResponse'
        "400":
          description: Invalid `before` or `limit`
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "403":
          description: Requester is not a participant (NOT_A_PARTICIPANT)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "404":
          description: Meeting not found
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List chat messages for a meeting
      tags:
      - Meetings
  /api/v1/meetings/{code}/ws:
    get:
      description: Upgrades the connection to WebSocket and places the caller into
        the meeting hub. Callers authenticate by presenting a short-lived ticket obtained
        from POST /meetings/:code/ws-ticket; the ticket is single-use and is consumed
        atomically at upgrade time. Callers who are scope members join directly; others
        enter the waiting room (knock flow).
      parameters:
      - description: Meeting code
        example: abc-defg-hij
        in: path
        name: code
        required: true
        type: string
      - description: Short-lived WS ticket (see /ws-ticket)
        in: query
        name: ticket
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols — WebSocket connection established
          schema:
            type: string
        "401":
          description: Missing (TICKET_REQUIRED) or invalid (TICKET_INVALID) or code-mismatched
            (TICKET_MISMATCH) ticket
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "403":
          description: Access denied (meeting ended or at capacity)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "404":
          description: Meeting not found
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      summary: Join meeting via WebSocket
      tags:
      - Meetings
  /api/v1/meetings/{code}/ws-ticket:
    post:
      description: Exchanges the caller's bearer token for a short-lived (60s), single-use
        ticket that authenticates the meeting WebSocket handshake. The ticket is bound
        to the calling user and the meeting code. Returns the ticket value and a fully-qualified
        ws_url path.
      parameters:
      - description: Meeting code
        example: abc-defg-hij
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Ticket issued
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.WSTicketResponse'
        "401":
          description: Missing or invalid bearer token
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "404":
          description: Meeting not found (MEETING_NOT_FOUND)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "410":
          description: Meeting has ended (MEETING_ENDED)
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Issue a WebSocket ticket for a meeting
      tags:
      - Meetings
  /api/v1/meetings/mine:
    get:
      description: Returns all meetings the authenticated user hosted or participated
        in, enriched with computed duration and up to 3 participant previews. Supports
        filtering by status and sorting by multiple criteria.
      parameters:
      - description: Filter by status
        enum:
        - in_progress
        - complete
        - processing
        - failed
        in: query
        name: filter[status]
        type: string
      - description: Filter by scope
        enum:
        - organization
        - department
        - open
        in: query
        name: filter[scope_type]
        type: string
      - description: UUID of the org or dept; required when scope_type is organization
          or department
        in: query
        name: filter[scope_id]
        type: string
      - default: created_at_desc
        description: Sort order
        enum:
        - created_at_desc
        - created_at_asc
        - duration_desc
        - duration_asc
        - title_asc
        - title_desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of meetings
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.MeetingListResponse'
        "400":
          description: Malformed scope parameters
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "403":
          description: Caller is not a member of the requested scope
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_rekall_backend_internal_interfaces_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my meetings
      tags:
      - Meetings
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT access token as: **Bearer &lt;token&gt;**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"